agent/requirements.txt
google-adk>=0.3.0
google-genai>=0.3.0
python-dotenv>=1.0.0
protobuf>=4.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.27.0
asyncpg>=0.29.0
aiosqlite>=0.19.0
